+++ DomainModels/Dto/RoomTypePutDto.cs
@@ -2,6 +2,6 @@ public class RoomTypePutDto
 {
     public int Id { get; set; }
     public string? Description { get; set; }
-    public double? Price { get; set; }
+    public decimal? Price { get; set; }
     public DateTime UpdatedAt { get; set; }
 }
